requirements.txt
timm
torch
torchvision
numpy
pandas
scikit-learn
pillow
tqdm
SimpleITK

